requirements.txt
torch==2.5.1+cu118
torchvision==0.20.1+cu118
scikit-learn==1.6.1
scipy==1.12.0
shimmy==2.0.0
stable-baselines3==2.5.0
pandas==2.2.3
numpy==1.26.3
empyrical==0.5.5
matplotlib==3.7.0
stockstats==0.6.4
gymnasium==1.0.0
ta-lib==0.5.2
tensorboard==2.18.0
tensorboardX==2.6.2.2
universal-portfolios==0.4.12
